+++ src/main/kotlin/oggvik/mods/configurablezombieai/runtime/abnormals/targetacquisition/AbnormalTargetAcquisitionCandidate.kt
@@ -1,0 +1,8 @@
+package oggvik.mods.configurablezombieai.runtime.abnormals.targetacquisition
+
+import net.minecraft.entity.LivingEntity
+
+data class AbnormalTargetAcquisitionCandidate(
+    val target: LivingEntity,
+    val distance: Double
+)